requirements.txt
# Core Django dependencies
Django>=4.2.0,<5.0
django-environ>=0.11.0
psycopg2-binary>=2.9.0

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.0

# Celery for async tasks
celery>=5.3.0
redis>=5.0.0
django-redis>=5.4.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# File handling and storage
Pillow>=10.0.0
django-storages>=1.14.0
boto3>=1.28.0  # For AWS S3 in production

# Authentication and permissions
django-guardian>=2.4.0
django-allauth>=0.57.0

# Development and testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=19.0.0

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
pre-commit>=3.4.0

# Documentation
django-extensions>=3.2.0
django-debug-toolbar>=4.2.0

# Utilities
python-dateutil>=2.8.0
requests>=2.31.0
openpyxl>=3.1.0  # For Excel export
reportlab>=4.0.0  # For PDF generation
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7

# Monitoring and logging
sentry-sdk>=1.32.0
django-health-check>=3.17.0